requirements.txt
Django~=5.1
qrcode~=7.4.2
djangorestframework~=3.15.2